use default request pagination values + minor refactor
