+++ activiti-cloud-audit-service/activiti-cloud-services-audit-jpa/src/main/java/org/activiti/cloud/services/audit/jpa/controllers/csv/ObjectToJsonConvertor.java
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2017-2020 Alfresco Software, Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.activiti.cloud.services.audit.jpa.controllers.csv;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.opencsv.bean.AbstractBeanField;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class ObjectToJsonConvertor extends AbstractBeanField {
+
+    private static final Logger logger = LoggerFactory.getLogger(ObjectToJsonConvertor.class);
+
+    private ObjectMapper objectMapper;
+
+    public ObjectToJsonConvertor(ObjectMapper objectMapper) {
+        this.objectMapper = objectMapper;
+    }
+
+    @Override
+    protected Object convert(String s) {
+        return null;
+    }
+
+    @Override
+    public String convertToWrite(Object value) {
+        try {
+            return objectMapper.writeValueAsString(value);
+        } catch (JsonProcessingException e) {
+            logger.error("Could serialise event " + value + " to json in cvs export.");
+        }
+        return "";
+    }
+
+
+}
